+++ events/js/events.client.js
@@ -1,0 +1,3 @@
+app.validUser().catch(()=>{
+  app.redirectWithPreloader('../register')
+})